Move prob to src/pyhf
